Extract the subheader logic into the PlansPageSubheader component
